test/smoke: extract Flags and Bin
- In order to share common methods between client and server.
